+++ test/test_url_open.py
@@ -531,5 +531,5 @@ def test_walk():
 def test_ssh_params():
 	with url.Context():
 		u = url.URL("ssh://livpython@www.livinglogic.de/~/checkouts/LivingLogic.Python.xist/")
-		assert u.isdir(remotepython="/usr/local/bin/python3.2") is True
+		assert u.isdir(python="/usr/local/bin/python3.2") is True
 		assert u.isdir(nice=20) is True